requirements.txt
# Essential packages only
# pandas==2.0.3
# numpy==1.24.3
# matplotlib==3.7.2
# seaborn==0.12.2
# scikit-learn==1.3.0
# xgboost==1.7.6
# joblib==1.3.1
# streamlit==1.25.0
# flask==2.3.2
# plotly==5.15.0
# jupyter==1.0.0

# Essential packages only
streamlit==1.37.1
numpy==2.1.3
pandas==2.2.3
matplotlib==3.9.2
seaborn==0.13.2
plotly==5.16.1
scikit-learn==1.5.2
xgboost==2.1.2
joblib==1.4.2
flask==2.3.2
scikit-surprise==1.1.3